backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary
PyJWT==2.6.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.3
